Auth-Service/requirements.txt
flask==1.1.2
gunicorn==19.8.1
google-cloud-firestore==1.7.0
pyjwt